+++ src/ui/react/settings/modules/shared/store/app/StateBridge.ts
@@ -20,6 +20,7 @@ export const StateToJsonSettings = (state: RootState): Settings => {
     "devTools",
     "language",
     "dateFormat",
+    "startOfWeek",
     "updater",
     "byWidget",
     "drpc",